+++ react/components/ProductImages/components/ThumbnailSlider.js
@@ -18,15 +18,11 @@ class ThumbnailSlider extends Component {
    * Function that configure slider settings according to the component props
    */
   configureSliderSettings = () => {
-    const { images, maxVisibleItems, orientation } = this.props
+    const { maxVisibleItems, orientation } = this.props
     const sliderVertical = orientation === VERTICAL
 
-    /** Number of visible slider items should not exceed the number of images, in the case that this condition is not satisfied,
-     * then it must not exceed the maxVisibleItems property. Finally, the number of visible slider items should not be greater
-     * than the maximum allowed value that is defined by MAX_VISIBLE_ITEMS */
     const numOfVisibleItems = Math.min(
       maxVisibleItems,
-      images.length,
       MAX_VISIBLE_ITEMS
     )